separated out tests (travis does not have latex!)
