+++ test/tools_plotting.py
@@ -3,8 +3,7 @@ Created on 5 May 2015
 
 @author: kreczko
 '''
-from tools.plotting import get_best_max_y, Histogram_properties,\
-    make_shape_comparison_plot, make_control_region_comparison
+from tools.plotting import get_best_max_y
 from tools.hist_utilities import value_errors_tuplelist_to_graph
 from tools.hist_utilities import value_error_tuplelist_to_hist
 
@@ -33,28 +32,3 @@ def test_get_max_y_hist():
 def test_get_max_y_graph():
     max_y = get_best_max_y([g1])
     assert max_y == 3 + 1
-
-def test_make_control_region_comparison():
-    # not a unit test
-    hp = Histogram_properties()
-    hp.name = 'make_control_region_comparison'
-    hp.mc_error = 0.0
-    make_control_region_comparison(
-        control_region_1 = h1,
-        control_region_2 = h2,
-        name_region_1 = 'h1',
-        name_region_2 = 'h2',
-        histogram_properties = hp,
-        save_folder = 'test/plots' )
-
-def test_make_shape_comparison_plot():
-    # not a unit test
-    hp = Histogram_properties()
-    hp.name = 'make_shape_comparison_plot'
-    make_shape_comparison_plot(
-        shapes = [h1, h2],
-        names = ['h1', 'h2'],
-        colours = ['yellow', 'red'],
-        histogram_properties = hp,
-        alpha= 1.0,
-        save_folder = 'test/plots')
